switch to CMake based build for neon subdirectory

+++ neon/benchmark_driver.cpp
@@ -97,7 +97,7 @@ int main() {
     int repetitions = 4000;
 
     // Evaluate each kernel.
-    for (int i = 0; i < kernels_count; i++) {
+    for (size_t i = 0; i < kernels_count; i++) {
         warmup_kernel(kernels[i], repetitions);
         handle_kernel_total(kernels[i], repetitions);
     }
